moved util out of core
