+++ src/config/factories/services/crud.ts
@@ -1,10 +1,10 @@
 import { resolve } from 'path';
 import { RequireSome } from '../../../core/interface';
-import { readDir } from '../../../core/util/file/readDir';
-import { readJsonFile } from '../../../core/util/file/readFile';
 import { crudService, CrudServiceProperties, Record } from '../../../middleware/services/crud.service';
 import { ServiceFactory } from '../../interfaces';
 import { $Env, Environment } from '../../utils/environment';
+import { readDir } from '../../utils/file/readDir';
+import { readJsonFile } from '../../utils/file/readFile';
 
 export interface CrudConfig extends RequireSome<CrudServiceProperties, 'route'> {
 	/** load crud data from disk */
