+++ src/middleware/util/file/getStat.ts
